@@ -1,20 +1,15 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
 import Header from './components/Header'
-import Content from './components/Content'
+import Editor from './components/Editor'
 import Footer from './components/Footer'
 import SaveDialog from './components/SaveDialog'
-import { fetchData, editBlock, openFile } from './actions'
+import { fetchData, openFile } from './actions'
 import { editorSelector } from './selectors'
 
 class Notebook extends Component {
-
-  constructor (props) {
-    super(props)
-    this.deselectBlocks = this.deselectBlocks.bind(this)
-  }
-
   componentWillMount () {
     this.props.dispatch(openFile())
   }
@@ -23,18 +18,12 @@ class Notebook extends Component {
     this.props.dispatch(fetchData())
   }
 
-  deselectBlocks () {
-    this.props.dispatch(editBlock(null))
-  }
-
   render () {
-    const { editable, saving, activeBlock } = this.props
-    const cssClass = editable ? ' editable' : ''
+    const { saving } = this.props
     const notebookView = (
-      <div className={'pure-u-1 pure-u-md-3-4 pure-u-lg-2-3' + cssClass}>
-        <Header editable={editable} />
-        <hr className='top-sep'></hr>
-        <Content editable={editable} activeBlock={activeBlock} />
+      <div className='pure-u-1 pure-u-md-3-4 pure-u-lg-2-3'>
+        <Header />
+        <Editor />
         <Footer />
       </div>
     )
@@ -45,7 +34,7 @@ class Notebook extends Component {
     )
     const content = saving ? saveView : notebookView
     return (
-      <div className='pure-g' onClick={this.deselectBlocks}>
+      <div className='pure-g'>
         <div className='offset-col pure-u-1 pure-u-md-1-8 pure-u-lg-1-6'>
           &nbsp;
         </div>
@@ -53,14 +42,11 @@ class Notebook extends Component {
       </div>
     )
   }
-
 }
 
 Notebook.propTypes = {
-  activeBlock: React.PropTypes.string,
-  saving: React.PropTypes.bool,
-  editable: React.PropTypes.bool,
-  dispatch: React.PropTypes.func
+  saving: PropTypes.bool,
+  dispatch: PropTypes.func
 }
 
 export default connect(editorSelector)(Notebook)
