@@ -4,16 +4,25 @@ import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import { createStore, compose, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
+import {enableBatching} from 'redux-batched-actions'
 import bindStoreToMenu from './bindStoreToMenu'
 
-import NotebookReducer from './reducers'
+import reducer from './reducers'
 import Notebook from './Notebook'
 
 require('../scss/main.scss')
 
+const middlewares = [
+  thunk
+]
+
+if (process.env.NODE_ENV !== 'production') {
+  middlewares.push(require('redux-logger').default)
+}
+
 const store = compose(
-    applyMiddleware(thunk)
-)(createStore)(NotebookReducer)
+    applyMiddleware(...middlewares)
+)(createStore)(enableBatching(reducer))
 
 bindStoreToMenu(store)
 
