@@ -4,7 +4,7 @@ const ExternalsPlugin = webpack.ExternalsPlugin
 const WriteFilePlugin = require('write-file-webpack-plugin')
 
 module.exports = {
-  devtool: 'eval-source-map',
+  devtool: 'cheap-module-source-map',
   entry: [
     'webpack-dev-server/client?http://localhost:3002',
     'webpack/hot/only-dev-server',
@@ -27,10 +27,10 @@ module.exports = {
       { test: /\.css$/, loader: 'style-loader!css-loader' },
       {
         test: /\.scss$/,
-        loaders: ['style', 'css', 'sass']
+        loaders: ['style-loader', 'css-loader', 'sass-loader']
       },
       { test: /\.png$/, loader: 'url-loader?limit=100000' },
-      { test: /\.jpg$/, loader: 'file-loader' },
+      { test: /\.(jpg|gif|png)$/, loader: 'file-loader' },
       { test: /\.json$/, loader: 'json-loader' },
       {
         test: /\.(ttf|eot|svg|otf|woff(2)?)(\?[a-z0-9=&.]+)?$/, // font files
@@ -38,7 +38,7 @@ module.exports = {
       },
       {
         test: /\.js$/,
-        loaders: ['react-hot', 'babel'],
+        loaders: ['react-hot-loader', 'babel-loader'],
         include: path.join(__dirname, 'src')
       }
     ]
